FizzBuzz solution in a Rails app :train:
Using rails for this simple UI is an overkill, but was a funny
experimenting this way.

For people that is learning Rails I think it's a
meaningfully exercise, since been able to interact with any solution you
create is always exciting.

This case we just created a UI for our FizzBuzz kata solution

The most important components of the integration of the FizzBuzz
functionality are:

- app/controllers/fizz_buzz_controller.rb
- app/controllers/fizz_buzz/list_controller.rb
- app/views/fizz_buzz/*
- app/models/fizz_buzz.rb
- spec/models/fizz_buzz_spec.rb
- spec/features/fizz_buzz_list_spec.rb